Extract simple settings sections
